* wav-pipe:
  README.adoc: fix typo
  README.adoc: update raw streams description (wav-pipe format related)
  README.adoc: basic documentation for wav-pipe format
  Bump copyright date in wav-pipe format.
  Check wav input format type from header.
  Use more efficient float->int conversion in SFOutputStream.
  Bump samples per read/write buffer size for wav pipe I/O.
  Improve wav-pipe format input error handling.
  Use similar API for from/to raw conversion in RawConverter.
  WavPipeInputStream: avoid filling sample array with zeros before read.
  Selectively enable -O3 for gcc for some functions (use auto vectorization).
  Optimize wav-pipe format a bit more.
  Optimize native endian case for RawConverter float->int 16/32 bit.
  Support 32-bit output pipe wav.
  Optimize RawConverter native endian raw->float conversion.
  BUILD: add endianness test
  Support new format (--format wav-pipe) for wav pipes without size limit.
  Support 32 bit conversions in RawConverter.

+++ src/sfinputstream.cc
@@ -155,7 +155,7 @@ SFInputStream::read_frames (vector<float>& samples, size_t count)
        * used during read and write are identical
        */
       samples.resize (r_count * m_n_channels);
-      const double norm = 1.0 / 0x80000000LL;
+      const float norm = 1.0 / 0x80000000LL;
       for (size_t i = 0; i < samples.size(); i++)
         samples[i] = isamples[i] * norm;
     }
